Fixing a deprecated behaviour for new versions of django
This warning indicates the need to remove the wrapping Context() call
in the code and just pass in a dictionary.

/usr/local/share/WAM/loads/management/commands/email_reminders.py:124:
RemovedInDjango110Warning: render() must be called with a dict, not a
Context.
  text_content = plaintext.render(context)

/usr/lib/python3/dist-packages/django/template/defaulttags.py:208:
RemovedInDjango110Warning: Need 2 values to unpack in for loop; got 3.
This will raise an exception in Django 1.10.
  RemovedInDjango110Warning)

/usr/local/share/WAM/loads/management/commands/email_reminders.py:125:
RemovedInDjango110Warning: render() must be called with a dict, not a
Context.
  html_content = html.render(context)